# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hooshyar AI Chat Platform API
  version: 1.0.0
  description: |
    API contract for the multi-model AI chat platform MVP (Persian-first, RTL).

    **Base conventions**

    * All timestamps are ISO-8601 UTC strings (`format: date-time`).
    * All ids are UUID v4 strings.
    * All endpoints require a Bearer JWT **except** `register` and `login`.
    * JWTs expire after 24 hours. When expired or invalid, every protected
      endpoint returns `401 UNAUTHENTICATED`.

    **Error contract**

    Every non-2xx JSON response uses one envelope:

    ```json
    { "error": { "code": "RESOURCE_NOT_FOUND", "message": "گفتگو پیدا نشد." } }
    ```

    `code` is machine-readable (stable, for frontend branching); `message` is a
    Persian, user-displayable string. Provider internals (names, hosts, HTTP
    codes, timeouts) never appear in any response.

    **Streaming (POST /conversations/{id}/messages)**

    Success responds `200` with `text/event-stream` (SSE over POST because
    EventSource cannot send an Authorization header or a body). Pre-stream
    validation failures (auth, ownership, model resolution, message validation)
    return the standard JSON error responses below — **not** SSE events. Once
    the stream opens, exactly one terminal event (`done` or `error`) is emitted:

    ```
    event: meta   data: {"userMessage": {...}, "model": {...}}   # once, first
    event: delta  data: {"text": "chunk"}                        # 0..n times
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /conversations:
    get:
      tags: [Conversations]
      operationId: listConversations
      summary: List the caller's conversations
      description: Returns only conversations owned by the authenticated user (INV-1), newest first. Unpaginated by design.
      responses:
        '200':
          description: Conversation collection (may be empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Conversations]
      operationId: createConversation
      summary: Create a conversation
      description: Creates a conversation owned by the caller. Omit the body (or send `{}`) to use the default title.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /conversations/{conversationId}/messages:
    post:
      tags: [Chat]
      operationId: sendMessage
      summary: Send a message and stream the AI answer (SSE)
      description: |
        Persists the user message, resolves the model (explicit `modelId`
        when provided and active, otherwise the default model), calls the AI
        provider, and streams the answer back as Server-Sent Events.

        **Order of checks (all pre-stream, returning normal JSON errors):**
        1. auth → `401`
        2. UUID shape → `400 VALIDATION_ERROR`
        3. message validation → `400 VALIDATION_ERROR`
        4. conversation exists **and** belongs to caller → `404 RESOURCE_NOT_FOUND`
        5. model resolution: unknown id → `404`; inactive id → `400 MODEL_INACTIVE`;
           no default configured → `400 NO_DEFAULT_MODEL`

        Then the stream opens (see the Streaming section in the API info):
        `meta` → `delta`* → `done` | `failed`. Exactly one assistant message is
        persisted per turn in every outcome (INV-5). The provider request is
        aborted after the configured timeout (A-8, default 60s) and surfaces as
        the generic `failed` terminal event.

        The generation runs DETACHED from this HTTP response: a client
        disconnect (refresh, closed tab, network loss) only unsubscribes — the
        AI keeps generating and the row completes in the DB. A returning
        client re-attaches via
        `/conversations/{conversationId}/messages/{messageId}/stream`.

        The first message of an untitled conversation auto-sets its title (A-7).
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              defaultModel:
                value: { content: 'سلام، یک ایده برای پروژه بده.' }
              explicitModel:
                value:
                  content: 'Explain this code.'
                  modelId: '9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d'
      responses:
        '200':
          description: |
            SSE stream (`text/event-stream`). Event payloads use the
            `StreamMetaEvent`, `StreamDeltaEvent`, `StreamDoneEvent`, and
            `StreamFailedEvent` schemas documented in components.

            Disconnect handling: disconnect ≠ failure. The generation is
            detached from the HTTP connection, so a mid-stream disconnect only
            unsubscribes the client — the generation continues, keeps
            persisting progress, and finishes `completed`. No error event is
            emitted to the disconnector (it cannot receive one). Provider /
            network failures persist with `status = "failed"` and emit a
            generic `StreamFailedEvent`. See
            `docs/CONVERSATION_RESILIENCE.md` §5 for the disambiguation.
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: meta
                data: {"userMessage":{"id":"e083f38c-155a-4ada-a85a-e04191d71831","conversationId":"3f7a9529-fcc2-4d6d-9723-ac6d69f928af","role":"user","content":"سلام","status":null,"errorMessage":null,"modelId":null,"clientMessageId":"cm-l8x9k-3a7b","createdAt":"2026-09-14T13:38:01.304Z"},"assistantMessage":{"id":"17d53156-1e46-4c61-b444-3ba9506b2d0b","conversationId":"3f7a9529-fcc2-4d6d-9723-ac6d69f928af","role":"assistant","content":"","status":"pending","errorMessage":null,"modelId":"84bc6ca9-b6c2-485f-828f-a48c32524759","clientMessageId":null,"createdAt":"2026-09-14T13:38:01.310Z"},"model":{"id":"84bc6ca9-b6c2-485f-828f-a48c32524759","name":"Mock GPT","provider":"mock"},"replay":false}

                event: delta
                data: {"text":"سلام! "}

                event: delta
                data: {"text":"این یک پاسخ آزمایشی است."}

                event: done
                data: {"assistantMessage":{"id":"17d53156-1e46-4c61-b444-3ba9506b2d0b","conversationId":"3f7a9529-fcc2-4d6d-9723-ac6d69f928af","role":"assistant","content":"سلام! این یک پاسخ آزمایشی است.","status":"completed","errorMessage":null,"modelId":"84bc6ca9-b6c2-485f-828f-a48c32524759","clientMessageId":null,"createdAt":"2026-09-14T13:38:03.352Z"}}
        '400':
          description: 'Validation failure, inactive model, or no default model. Codes: `VALIDATION_ERROR`, `MODEL_INACTIVE`, `NO_DEFAULT_MODEL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Unknown/foreign conversation or unknown model id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations/{conversationId}/messages/{messageId}/stream:
    get:
      tags: [Chat]
      operationId: reconnectGenerationStream
      summary: Reconnect to an assistant message's generation (SSE recovery stream)
      description: |
        Recovery stream for a client that lost its live connection — refresh
        mid-stream, closed tab, restored network, or a second tab opening the
        same conversation. Never re-invokes the AI:

        - live generation → `StreamSnapshotEvent` + remaining
          `StreamDeltaEvent`s + `StreamDoneEvent`/`StreamFailedEvent`
        - `completed` row → snapshot + done (pure replay)
        - `failed`/`interrupted` row → snapshot + `StreamFailedEvent`
        - `pending`/`streaming` row with NO live generation (server restarted
          mid-generation) → the row is honestly marked `interrupted` in the DB,
          then snapshot + `StreamFailedEvent`; the client can retry

        Subscribe-first/snapshot-second ordering guarantees no token is lost
        or duplicated: snapshot + deltas concatenate exactly to the persisted
        content.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: |
            SSE stream (`text/event-stream`): `snapshot` → `delta`* →
            `done` | `failed`.
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: snapshot
                data: {"assistantMessage":{"id":"17d53156-1e46-4c61-b444-3ba9506b2d0b","role":"assistant","content":"سلام! این یک","status":"streaming","errorMessage":null}}

                event: delta
                data: {"text":"پاسخ آزمایشی است."}

                event: done
                data: {"assistantMessage":{"id":"17d53156-1e46-4c61-b444-3ba9506b2d0b","role":"assistant","content":"سلام! این یک پاسخ آزمایشی است.","status":"completed","errorMessage":null}}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Unknown/foreign conversation, or the message is not an assistant row of it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer <accessToken> from register/login. Expires in 24h.'
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation UUID.
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional idempotency token. Mirror of `clientMessageId` on the request
        body — the backend accepts either, with the body field taking precedence
        when both are present. See `SendMessageRequest.clientMessageId` for the
        replay / 400 semantics.
      schema:
        type: string
        maxLength: 64
  responses:
    BadRequest:
      description: 'Invalid input (shape, format, blank content, bad UUID, etc.). Code: `VALIDATION_ERROR`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: 'متن پیام نمی‌تواند خالی یا فقط فاصله باشد.'
    Unauthorized:
      description: 'Missing, malformed, or expired token. Code: `UNAUTHENTICATED`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: 'توکن نامعتبر یا منقضی شده است.'
...
    InternalError:
      description: 'Unexpected server error; details are logged server-side only. Code: `INTERNAL_ERROR`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: 'خطای غیرمنتظره‌ای رخ داد. لطفاً دوباره تلاش کنید.'
  schemas:
...
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Optional; defaults to «گفتگوی جدید». Auto-set from the first message when left default (A-7).
    Conversation:
      type: object
      required: [id, title, userId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        userId:
          type: string
          format: uuid
          description: Owning user (INV — exactly one owner per conversation).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          description: Bumped on new messages; used for newest-first ordering.
...
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
          pattern: \S
          description: Non-blank after trim (BR-4); `pattern` requires at least one non-whitespace character.
        modelId:
          type: string
          format: uuid
          description: |
            Optional; must reference a model that is active and allowed for
            the caller's plan (active + free for FREE-plan users). Omit to
            use the default model. Authorization is re-verified on every
            send; a premium model requested by a FREE user is rejected with
            403 even via direct API calls.
        clientMessageId:
          type: string
          maxLength: 64
          description: |
            Optional idempotency token. The same value may also be sent as
            the `Idempotency-Key` HTTP header — both are accepted.
              * No matching row → new user message persisted with this id.
              * Match + matching content → replay (same user row, fresh
                assistant row, `meta.replay = true`).
              * Match + mismatched content → 400 in pre-flight
                (`این پیام قبلاً با متن دیگری ارسال شده است.`).

    # ---- Models ----
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - EMAIL_ALREADY_EXISTS
                - MODEL_INACTIVE
                - NO_DEFAULT_MODEL
                - DEFAULT_MODEL_PROTECTED
                - UNAUTHENTICATED
                - FORBIDDEN
                - RESOURCE_NOT_FOUND
                - INTERNAL_ERROR
              description: |
                Stable machine-readable code. Status-code mapping:
                400 → VALIDATION_ERROR | MODEL_INACTIVE | NO_DEFAULT_MODEL | DEFAULT_MODEL_PROTECTED;
                401 → UNAUTHENTICATED; 403 → FORBIDDEN; 404 → RESOURCE_NOT_FOUND;
                409 → EMAIL_ALREADY_EXISTS; 500 → INTERNAL_ERROR.
            message:
              type: string
              description: Persian, user-displayable message (safe; no technical detail).